[backend] add lock to loadDraftElement operation
